codegen: Add go:generate to command line in output
The generated output for Go source contains the command line used to
generate it, as a comment. Turn that comment into a `//go:generate`
comment so the file can be regenerated by running `go generate`.

Also add a blank line before the `package` clause so that godoc will not
treat the boilerplate as documentation.