codegen: Don't strip space/nl from nodes
Do not strip spaces or newlines from nodes in the AST when generating Go
code. If there is a space or newline in the AST, then it should be there
in the generated code. The codegen should just make sure that the
emitted code will compile and do whatever escaping is necessary to
ensure that.

Issues: #59